requirements.txt
# ─── Core ML ───────────────────────────────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
scipy>=1.11.0

# ─── Pose Estimation ───────────────────────────────────────────────────────────
# Install mmcv first: pip install mmcv==2.1.0 -f https://download.openmmlab.com/mmcv/dist/{cu121}/torch2.1/index.html
mmengine>=0.10.0
mmdet>=3.2.0
mmpose>=1.3.0

# ─── Web Backend ───────────────────────────────────────────────────────────────
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-multipart>=0.0.9
python-dotenv>=1.0.0

# ─── LLM ───────────────────────────────────────────────────────────────────────
google-generativeai>=0.4.0

# ─── Vision / Utilities ────────────────────────────────────────────────────────
opencv-python>=4.9.0
Pillow>=10.2.0
tqdm>=4.66.0

# ─── Training (also needed in Colab) ───────────────────────────────────────────
gdown>=5.1.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
tensorboard>=2.16.0
